rename package to env0
