+++ env0/resource_configuration_variable.go
@@ -1,4 +1,4 @@
-package env0tfprovider
+package env0
 
 import (
 	"context"
