+++ main.go
@@ -1,12 +1,12 @@
 package main
 
 import (
-	"github.com/env0/terraform-provider-env0/env0tfprovider"
+	"github.com/env0/terraform-provider-env0/env0"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/plugin"
 )
 
 func main() {
 	plugin.Serve(&plugin.ServeOpts{
-		ProviderFunc: env0tfprovider.Provider,
+		ProviderFunc: env0.Provider,
 	})
 }
